requirements.txt
# Chaptrix dependencies - Updated for Unified Workflow

# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
lxml>=4.9.0

# Image processing
opencv-python>=4.6.0
numpy>=1.22.0

# Utilities
colorama>=0.4.6  # New dependency for config validator
tqdm>=4.64.0     # New dependency for progress bars

# Google Drive integration
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0

# Discord integration
discord.py>=2.1.0
python-dotenv>=0.19.0

# Web dashboard (optional)
streamlit>=1.18.0